[local] Fix the load minigraph failed issue using a new method
If load minigraph before some critical services are started, it may
fail with error like 'service swss.service is cancelled'. Simply
checking state of syncd or swss service using systemctl is not enough.

According to the design of the services, a process like 'docker wait
<service_name>' is started in the final stage of starting service.
This fix is to wait until such process of all critical services are
started, then run load minigraph.

Signed-off-by: Xin Wang <xinw@mellanox.com>