to fix bug 1717512, fdb test fails sporadically. we have investigated it and found that the failure is because the DUT fails to learn MAC address and then flood the packet in the vlan. (#928)
The working flow of fdb test contains two steps:
1. PTF docker sends the learning frame with a view to having the DUT learnt the source mac address of the frame.
2. PTF docker sends the test frame and check whether the test frame is received from the port where the learning frame sent.
Typecally the DUT require a short period to have the MAC address learnt. However, there is no delay between step 1 and 2 above. There is a probability of 20% that when the DUT hasn't learnt the MAC address when it receives the test frame. In that case the test frame will be flooded.
We introduce a short delay of 2 seconds between the above steps 1 and 2 in order to solve that problem and this issue doesn't occur with the delay.