[local] align port_config.ini and sai_2010.xml for SN2010
There appeared an issue that initial and desired port mappings do not match.
This is crucial for warm reboot and most recent SDK will crash imidiately while
before it was silently making disruptive changes.
So, this is a potential fix for > 1 sec. downtime [RM https://redmine.mellanox.com/issues/1851074]

The reason why port_config.ini was changed as well is because it is not possible to configure
4x port into 2x1 ports and 2 lanes unmapped from SAI xml.
Luckly, two last ports become UP with 2x2 configuration.

Change-Id: I67bf7aaf98735fad06868add32ee7a477eb05d75
Signed-off-by: Stepan Blyschak <stepanb@mellanox.com>