[local] Fix the issue of duplicated psu status returned by get_psu_status
Duplicated PSU status may be returned by the get_psu_status of the
Mellanox specific PSU controller. Reason is that logic of "port_oid in
output line" is used for matching status line. In case the port_oid
ends with "1.1", the logic would match lines end with "1.1" and "1.11",
"1.12", etc.

The fix is to split the status line and use "==" for matching port_oid.

Change-Id: Ibea838d51c87674cc28a04232de87784a2a21167