[platform]Add teardown for reboot (#1395)
* [platform/test_reboot]Introduce the teardown module in which the all services, interfaces and transceivers are checked.
In some kinds of the reboot test the main purpose is to check the reboot cause.
Therefore these tests returns success without waiting other conditions to be true.
However, this fails the sanity check because other conditions aren't satisfied.
To avoid that, we introduce the teardown module to make sure all the necessary checkings passed ahead of exiting module

* fix syntax error

* fix typo

* Fix review comments

* fix comments

* [platform]Use a unified function to perform sanity check and clearify log info

Change-Id: Ibec96aef0922ab1577d957e1a8c9a742e6f06f5f